+++ src/Matcher/Dumper/StaticPrefixCollection.php
@@ -51,12 +51,12 @@ class StaticPrefixCollection
      */
     public function addRoute(string $prefix, $route): void
     {
-        list($prefix, $staticPrefix) = $this->getCommonPrefix($prefix, $prefix);
+        [$prefix, $staticPrefix] = $this->getCommonPrefix($prefix, $prefix);
 
         for ($i = \count($this->items) - 1; 0 <= $i; --$i) {
             $item = $this->items[$i];
 
-            list($commonPrefix, $commonStaticPrefix) = $this->getCommonPrefix($prefix, $this->prefixes[$i]);
+            [$commonPrefix, $commonStaticPrefix] = $this->getCommonPrefix($prefix, $this->prefixes[$i]);
 
             if ($this->prefix === $commonPrefix) {
                 // the new route and a previous one have no common prefix, let's see if they are exclusive to each others
@@ -90,8 +90,8 @@ class StaticPrefixCollection
             } else {
                 // the new route and a previous one have a common prefix, let's merge them
                 $child = new self($commonPrefix);
-                list($child->prefixes[0], $child->staticPrefixes[0]) = $child->getCommonPrefix($this->prefixes[$i], $this->prefixes[$i]);
-                list($child->prefixes[1], $child->staticPrefixes[1]) = $child->getCommonPrefix($prefix, $prefix);
+                [$child->prefixes[0], $child->staticPrefixes[0]] = $child->getCommonPrefix($this->prefixes[$i], $this->prefixes[$i]);
+                [$child->prefixes[1], $child->staticPrefixes[1]] = $child->getCommonPrefix($prefix, $prefix);
                 $child->items = [$this->items[$i], $route];
 
                 $this->staticPrefixes[$i] = $commonStaticPrefix;
@@ -135,42 +135,53 @@ class StaticPrefixCollection
         $baseLength = \strlen($this->prefix);
         $end = min(\strlen($prefix), \strlen($anotherPrefix));
         $staticLength = null;
-        set_error_handler([__CLASS__, 'handleError']);
+        set_error_handler([self::class, 'handleError']);
 
         for ($i = $baseLength; $i < $end && $prefix[$i] === $anotherPrefix[$i]; ++$i) {
             if ('(' === $prefix[$i]) {
                 $staticLength = $staticLength ?? $i;
+
                 for ($j = 1 + $i, $n = 1; $j < $end && 0 < $n; ++$j) {
                     if ($prefix[$j] !== $anotherPrefix[$j]) {
                         break 2;
                     }
+
                     if ('(' === $prefix[$j]) {
                         ++$n;
                     } elseif (')' === $prefix[$j]) {
                         --$n;
                     } elseif ('\\' === $prefix[$j] && (++$j === $end || $prefix[$j] !== $anotherPrefix[$j])) {
                         --$j;
+
                         break;
                     }
                 }
+
                 if (0 < $n) {
                     break;
                 }
+
                 if (('?' === ($prefix[$j] ?? '') || '?' === ($anotherPrefix[$j] ?? '')) && ($prefix[$j] ?? '') !== ($anotherPrefix[$j] ?? '')) {
                     break;
                 }
+
                 $subPattern = substr($prefix, $i, $j - $i);
+
                 if ($prefix !== $anotherPrefix && !preg_match('/^\(\[[^\]]++\]\+\+\)$/', $subPattern) && !preg_match('{(?<!'.$subPattern.')}', '')) {
                     // sub-patterns of variable length are not considered as common prefixes because their greediness would break in-order matching
                     break;
                 }
+
                 $i = $j - 1;
             } elseif ('\\' === $prefix[$i] && (++$i === $end || $prefix[$i] !== $anotherPrefix[$i])) {
                 --$i;
+
                 break;
             }
         }
+
         restore_error_handler();
+
         if ($i < $end && 0b10 === (\ord($prefix[$i]) >> 6) && preg_match('//u', $prefix.' '.$anotherPrefix)) {
             do {
                 // Prevent cutting in the middle of an UTF-8 characters
@@ -181,8 +192,8 @@ class StaticPrefixCollection
         return [substr($prefix, 0, $i), substr($prefix, 0, $staticLength ?? $i)];
     }
 
-    public static function handleError(int $type, string $msg, string $file, int $line): bool
+    public static function handleError(int $type, string $msg): bool
     {
-        return 0 === strpos($msg, 'preg_match(): Compilation failed: lookbehind assertion is not fixed length');
+        return false !== strpos($msg, 'Compilation failed: lookbehind assertion is not fixed length');
     }
 }