+++ tests/Fixtures/dumper/compiled_url_matcher0.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * This file has been auto-generated
+ * by the PubSubRouterBundle.
+ */
+return [
+    [ // $staticRoutes
+    ],
+    [ // $regexpList
+    ],
+    [ // $dynamicRoutes
+    ],
+];