+++ tests/Matcher/MatcherTest.php
@@ -4,6 +4,7 @@ namespace Gos\Bundle\PubSubRouterBundle\Tests\Matcher;
 
 use Gos\Bundle\PubSubRouterBundle\Exception\ResourceNotFoundException;
 use Gos\Bundle\PubSubRouterBundle\Matcher\Matcher;
+use Gos\Bundle\PubSubRouterBundle\Matcher\MatcherInterface;
 use Gos\Bundle\PubSubRouterBundle\Router\Route;
 use Gos\Bundle\PubSubRouterBundle\Router\RouteCollection;
 use PHPUnit\Framework\TestCase;
@@ -15,6 +16,7 @@ class MatcherTest extends TestCase
         // test the patterns are matched and parameters are returned
         $collection = new RouteCollection();
         $collection->add('foo', $route = new Route('foo/{bar}', 'strlen'));
+        $route->compile();
         $matcher = $this->getMatcher($collection);
 
         try {
@@ -28,6 +30,7 @@ class MatcherTest extends TestCase
         // test that defaults are merged
         $collection = new RouteCollection();
         $collection->add('foo', $route = new Route('foo/{bar}', 'strlen', ['def' => 'test']));
+        $route->compile();
         $matcher = $this->getMatcher($collection);
 
         $this->assertEquals(['foo', $route, ['bar' => 'baz', 'def' => 'test']], $matcher->match('foo/baz'));
@@ -35,6 +38,7 @@ class MatcherTest extends TestCase
         // route with an optional variable as the first segment
         $collection = new RouteCollection();
         $collection->add('bar', $route = new Route('{bar}/foo', 'strlen', ['bar' => 'bar'], ['bar' => 'foo|bar']));
+        $route->compile();
         $matcher = $this->getMatcher($collection);
 
         $this->assertEquals(['bar', $route, ['bar' => 'bar']], $matcher->match('bar/foo'));
@@ -42,6 +46,7 @@ class MatcherTest extends TestCase
 
         $collection = new RouteCollection();
         $collection->add('bar', $route = new Route('{bar}', 'strlen', ['bar' => 'bar'], ['bar' => 'foo|bar']));
+        $route->compile();
         $matcher = $this->getMatcher($collection);
 
         $this->assertEquals(['bar', $route, ['bar' => 'foo']], $matcher->match('foo'));
@@ -50,6 +55,7 @@ class MatcherTest extends TestCase
         // route with only optional variables
         $collection = new RouteCollection();
         $collection->add('bar', $route = new Route('{foo}/{bar}', 'strlen', ['foo' => 'foo', 'bar' => 'bar']));
+        $route->compile();
         $matcher = $this->getMatcher($collection);
 
         $this->assertEquals(['bar', $route, ['foo' => 'foo', 'bar' => 'bar']], $matcher->match(''));
@@ -61,7 +67,7 @@ class MatcherTest extends TestCase
     {
         $collection = new RouteCollection();
         $collection->add('$péß^a|', $route = new Route('bar', 'strlen'));
-
+        $route->compile();
         $matcher = $this->getMatcher($collection);
         $this->assertEquals(['$péß^a|', $route, []], $matcher->match('bar'));
     }
@@ -70,9 +76,11 @@ class MatcherTest extends TestCase
     {
         $collection = new RouteCollection();
         $collection->add('foo', $route = new Route('foo', 'strlen'));
+        $route->compile();
 
         $collection1 = new RouteCollection();
         $collection1->add('foo', $route = new Route('foo1', 'strlen'));
+        $route->compile();
 
         $collection->addCollection($collection1);
 
@@ -272,7 +280,7 @@ class MatcherTest extends TestCase
         $this->assertEquals('a', $matcher->match('éo')[0]);
     }
 
-    protected function getMatcher(RouteCollection $routes): Matcher
+    protected function getMatcher(RouteCollection $routes): MatcherInterface
     {
         return new Matcher($routes);
     }